+++ src/keywords/maximum.rs
@@ -3,48 +3,58 @@ use crate::{
     compilation::{CompilationContext, JSONSchema},
     error::{error, no_error, CompilationError, ErrorIterator, ValidationError},
 };
+use num_cmp::NumCmp;
 use serde_json::{Map, Value};
 
-pub struct MaximumValidator {
+pub struct MaximumU64Validator {
+    limit: u64,
+}
+pub struct MaximumI64Validator {
+    limit: i64,
+}
+pub struct MaximumF64Validator {
     limit: f64,
 }
 
-impl MaximumValidator {
-    #[inline]
-    pub(crate) fn compile(schema: &Value) -> CompilationResult {
-        if let Value::Number(limit) = schema {
-            let limit = limit.as_f64().expect("Always valid");
-            return Ok(Box::new(MaximumValidator { limit }));
-        }
-        Err(CompilationError::SchemaError)
-    }
-}
+macro_rules! validate {
+    ($validator: ty) => {
+        impl Validate for $validator {
+            fn validate<'a>(
+                &self,
+                schema: &'a JSONSchema,
+                instance: &'a Value,
+            ) -> ErrorIterator<'a> {
+                if self.is_valid(schema, instance) {
+                    no_error()
+                } else {
+                    error(ValidationError::maximum(instance, self.limit as f64)) // do not cast
+                }
+            }
 
-impl Validate for MaximumValidator {
-    fn validate<'a>(&self, _: &'a JSONSchema, instance: &'a Value) -> ErrorIterator<'a> {
-        if let Value::Number(item) = instance {
-            let item = item.as_f64().expect("Always valid");
-            if item > self.limit {
-                return error(ValidationError::maximum(instance, self.limit));
+            fn is_valid(&self, _: &JSONSchema, instance: &Value) -> bool {
+                if let Value::Number(item) = instance {
+                    return if let Some(item) = item.as_u64() {
+                        !NumCmp::num_gt(item, self.limit)
+                    } else if let Some(item) = item.as_i64() {
+                        !NumCmp::num_gt(item, self.limit)
+                    } else {
+                        let item = item.as_f64().expect("Always valid");
+                        !NumCmp::num_gt(item, self.limit)
+                    };
+                }
+                true
+            }
+
+            fn name(&self) -> String {
+                format!("maximum: {}", self.limit)
             }
         }
-        no_error()
-    }
-
-    fn is_valid(&self, _: &JSONSchema, instance: &Value) -> bool {
-        if let Value::Number(item) = instance {
-            let item = item.as_f64().expect("Always valid");
-            if item > self.limit {
-                return false;
-            }
-        }
-        true
-    }
-
-    fn name(&self) -> String {
-        format!("maximum: {}", self.limit)
-    }
+    };
 }
+
+validate!(MaximumU64Validator);
+validate!(MaximumI64Validator);
+validate!(MaximumF64Validator);
 
 #[inline]
 pub fn compile(
@@ -52,5 +62,28 @@ pub fn compile(
     schema: &Value,
     _: &CompilationContext,
 ) -> Option<CompilationResult> {
-    Some(MaximumValidator::compile(schema))
+    if let Value::Number(limit) = schema {
+        return if let Some(limit) = limit.as_u64() {
+            Some(Ok(Box::new(MaximumU64Validator { limit })))
+        } else if let Some(limit) = limit.as_i64() {
+            Some(Ok(Box::new(MaximumI64Validator { limit })))
+        } else {
+            let limit = limit.as_f64().expect("Always valid");
+            Some(Ok(Box::new(MaximumF64Validator { limit })))
+        };
+    }
+    Some(Err(CompilationError::SchemaError))
+}
+
+#[cfg(test)]
+mod tests {
+    use crate::tests_util;
+    use serde_json::{json, Value};
+    use test_case::test_case;
+
+    #[test_case(json!({"maximum": 1u64 << 54}), json!(1u64 << 54 + 1))]
+    #[test_case(json!({"maximum": 1i64 << 54}), json!(1i64 << 54 + 1))]
+    fn is_not_valid(schema: Value, instance: Value) {
+        tests_util::is_not_valid(schema, instance)
+    }
 }